requirements.txt
google-adk>=1.18.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
rich>=13.8.0

